parking_v5/parking_system/requirements.txt
flask
opencv-python
pyserial
